# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a music subscription service server.",
        "title": "Music Subscription Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Login with email and password to receive a JWT access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate user and retrieve token",
                "parameters": [
                    {
                        "description": "Credentials payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_auth_handler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_features_auth_handler.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user in the system with email, password, and username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_auth_handler.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_features_users_model.User"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (Email or Username already taken)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/favorites/tracks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all tracks added to favorites by the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorites"
                ],
                "summary": "List favorites",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/music-service_internal_features_tracks_model.Track"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/favorites/tracks/{track_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a track by ID to the user's favorites list. FREE users can add up to 20 favorites.",
                "tags": [
                    "favorites"
                ],
                "summary": "Add track to favorites",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Track ID",
                        "name": "track_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created (Track added to favorites)"
                    },
                    "400": {
                        "description": "Invalid track ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Favorites limit exceeded)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (Track already in favorites)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a track by ID from the user's favorites list",
                "tags": [
                    "favorites"
                ],
                "summary": "Remove track from favorites",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Track ID",
                        "name": "track_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (Track successfully removed from favorites)"
                    },
                    "400": {
                        "description": "Invalid track ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Track not found in favorites",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_features_auth_handler.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_features_auth_handler.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "username": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
        "internal_features_auth_handler.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
...
        "music-service_internal_core_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "music-service_internal_features_tracks_model.Track": {
            "type": "object",
            "properties": {
                "album_id": {
                    "type": "integer"
                },
                "album_title": {
                    "type": "string"
                },
                "artist_id": {
                    "type": "integer"
                },
                "artist_name": {
                    "description": "Join fields",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "genre_id": {
                    "type": "integer"
                },
                "genre_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "music-service_internal_features_users_model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "subscription_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "ВАЖНО: вводите с префиксом \"Bearer \". Пример: Bearer eyJhbGci...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}